requirements.txt
fastapi
unicorn
sqlalchemy
pymysql
passlib[argon2]
python-jose[cryptography]
python-multipart
pydantic[email]